+++ rabbithole-endpoint-actix/Cargo.toml
@@ -12,26 +12,28 @@ categories = []
 license = "MIT"
 
 [dependencies]
-env_logger = "~0.7"
-log = "~0.4"
-lazy_static = "~1.4"
+env_logger = "0.7.1"
+log = "0.4.8"
+lazy_static = "1.4.0"
 
-futures = { version = "~0.3" }
+futures = "0.3.5"
+
 actix-web = "2.0.0"
+actix-rt = "1.0"
 
-url = "~1.7"
-http = "~0.1"
+url = "2.1.1"
+http = "0.2.1"
 
-serde = "~1.0"
-serde_json = "~1.0"
+serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0.53"
 uuid = "~0.8"
 async-trait = "~0.1"
 rand = "~0.7"
-config = "~0.9"
+config = "0.10.1"
 
-rabbithole = { path = "../rabbithole", version = "~0.3" }
-rabbithole-derive = { path = "../rabbithole-derive", version = "~0.3" }
+rabbithole = { path = "../rabbithole" }
+rabbithole-derive = { path = "../rabbithole-derive" }
 
 [dev-dependencies]
-actix-http-test = "=0.3.0-alpha.1"
-actix-http = "=0.3.0-alpha.1"
+actix-http-test = "1.0.0"
+actix-http = "1.0.1"
