+++ rabbithole-endpoint-actix/examples/manual_crud.rs
@@ -8,37 +8,40 @@ use actix_web::App;
 use actix_web::{middleware, web};
 use actix_web::{HttpResponse, HttpServer};
 
-use rabbithole_endpoint_actix::settings::ActixSettingsModel;
 use rabbithole_endpoint_actix::ActixSettings;
 
 use crate::service::dog::DogService;
 use crate::service::human::HumanService;
-use std::convert::TryInto;
+use config::{Config, File};
 
-fn main() -> std::io::Result<()> {
+#[actix_rt::main]
+async fn main() -> std::io::Result<()> {
     std::env::set_var("RUST_LOG", "info");
     env_logger::init();
 
-    let mut settings = config::Config::default();
-    settings.merge(config::File::with_name("config/actix.config.example.toml")).unwrap();
-    let settings: ActixSettingsModel = settings.try_into().unwrap();
-    let settings_port = settings.port;
+    let mut settings = Config::default();
+
+    settings.merge(File::with_name("config/actix.config.example.toml")).unwrap();
+
+    let actix_settings: ActixSettings = settings.clone().try_into().unwrap();
 
     let dog_service = DogService::new();
     let human_service = HumanService::new(dog_service.clone());
+
     HttpServer::new(move || {
         App::new()
             .data(dog_service.clone())
             .data(human_service.clone())
-            .data::<ActixSettings>(settings.clone().try_into().unwrap())
+            .data::<ActixSettings>(actix_settings.clone())
             .wrap(middleware::Logger::default())
             .service(
-                web::scope(&settings.path)
+                web::scope(&actix_settings.path)
                     .service(DogService::actix_service())
                     .service(HumanService::actix_service()),
             )
             .default_service(web::to(HttpResponse::NotFound))
     })
-    .bind(format!("[::]:{}", settings_port))?
+    .bind(format!("[::]:{:?}", settings.get::<u32>("port")))?
     .run()
+    .await
 }
