--- rabbithole-endpoint-actix/tests/v1_1/mod.rs
@@ -1,1 +1,0 @@
-pub mod common_test;
