+++ rabbithole/Cargo.toml
@@ -29,10 +29,10 @@ serde_derive = "~1.0"
 itertools = "~0.8"
 num-integer = "~0.1"
 
-http = "~0.1"
-url = "~2.1"
+http = "0.2.1"
+url = "2.1.1"
 percent-encoding = "~2.1"
-base64 = "~0.11"
+base64 = "0.12.1"
 
 uuid = { version = "~0.8", features = ["v4", "serde"] }
 rsql = "~0.4"
