+++ rabbithole/tests/helper.rs
@@ -1,4 +1,3 @@
-use std::error::Error;
 use std::fs::File;
 use std::io::prelude::*;
 use std::path::Path;
@@ -8,14 +7,14 @@ pub fn read_json_file(filename: &str) -> String {
     let display = path.display();
 
     let mut file = match File::open(&path) {
-        Err(why) => panic!("couldn't open {}: {}", display, Error::description(&why)),
+        Err(why) => panic!("couldn't open {}: {}", display, &why),
         Ok(file) => file,
     };
 
     let mut s = String::new();
 
     if let Err(why) = file.read_to_string(&mut s) {
-        panic!("couldn't read {}: {}", display, Error::description(&why));
+        panic!("couldn't read {}: {}", display, &why);
     };
 
     s
